Restructure OSIDB entities commands
