+++ griffon/commands/plugins/__init__.py
@@ -1,0 +1,3 @@
+"""
+griffon plugins
+"""
